Move lib/messages.js to lib/slack-messages.js
