+++ lib/post_data.js
@@ -18,7 +18,7 @@ limitations under the License.
 "use strict";
 
 var env = process.env,
-  messages = require('./messages.js'),
+  messages = require('./slack-messages.js'),
   util = require('util'),
   utils = require('./utils.js');
