+++ lib/slack-messages.js
